+++ addons/main/config.cpp
@@ -29,6 +29,25 @@ class CfgPatches {
             "uk3cb_baf_vehicles_MAN"
         };
         author = "johnb43";
+        url = "https://github.com/johnb432/Weapons-Balance-ACE";
+        VERSION_CONFIG;
+    };
+};
+
+class CfgMods {
+    class PREFIX {
+        name = "Weapons Balance - 3CB BAF";
+        author = "johnb43";
+        tooltipOwned = "Weapons Balance - 3CB BAF";
+        hideName = 0;
+        hidePicture = 0;
+        actionName = "Github";
+        action = "https://github.com/johnb432/Weapons-Balance-3CBBAF";
+        description = "A collection of balancing mods made by johnb43.";
+        overview = "A collection of balancing mods made by johnb43.";
+        picture = "\z\wb_3cbbaf\addons\main\ui\logo_weapons_balance.paa"; // http://getdrawings.com/get-drawing#gun-drawing-in-pencil-27.jpg, http://getdrawings.com/get-drawing#gun-drawing-in-pencil-17.png
+        logo = "\z\wb_3cbbaf\addons\main\ui\logo_weapons_balance.paa";
+        overviewPicture = "\z\wb_3cbbaf\addons\main\ui\logo_weapons_balance.paa";
     };
 };