+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Changelog for Weapons balance 13.11.2021
+
+1.1.0.1
+- Minor tweaks.
+- Added weapon attachments compatibility.
+
 # Changelog for Weapons balance 9.8.2021
 
 1.1.0.0